Add files via upload
dice game using html css and javascript I created a basic dice simulator which produces three outcomes player 1 wins, player 2 wins and draw depending on the dices rolled. I made this using HTML and CSS. This is the first project where I used javascript to work the functionality of the project. I had a chance to use variables to capture inforamtion. Also had a chance to use DOM manipulation to pick elements that needed changing. Also introduced if else statements to allow me to produce answers depending on the outcomes of the dices.